feat: Create separated TableInput component [BLOX-205]
